+++ src/os2datascanner/projects/admin/adminapp/management/commands/quickstart_dev.py
@@ -150,8 +150,7 @@ class Command(BaseCommand):
             do_last_modified_check=False,
             organization=org,
             schedule=recurrence,
-            rule=cpr,
-            contact_person=user
+            rule=cpr
         )
         if created:
             auth = Authentication(username=smb_user)
@@ -172,8 +171,7 @@ class Command(BaseCommand):
             organization=org,
             schedule=recurrence,
             download_sitemap=False,
-            rule=cpr,
-            contact_person=user
+            rule=cpr
         )
         if created:
             self.stdout.write(self.style.SUCCESS("Webscanner created successfully!"))