+++ src/os2datascanner/projects/admin/adminapp/views/dropboxscanner_views.py
@@ -44,7 +44,7 @@ class DropboxScannerCreate(ScannerCreate):
         'only_notify_superadmin',
         'rule',
         'organization',
-        'contact_person'
+        'contacts'
     ]
 
     def get_form(self, form_class=None):
@@ -76,7 +76,7 @@ class DropboxScannerUpdate(ScannerUpdate):
         'only_notify_superadmin',
         'rule',
         'organization',
-        'contact_person'
+        'contacts'
     ]
 
     def get_form(self, form_class=None):