+++ src/os2datascanner/projects/admin/adminapp/views/googledrivescanner_views.py
@@ -45,7 +45,7 @@ class GoogleDriveScannerCreate(ScannerCreate):
         'only_notify_superadmin',
         'rule',
         'organization',
-        'contact_person'
+        'contacts'
     ]
 
     def get_success_url(self):
@@ -69,7 +69,7 @@ class GoogleDriveScannerUpdate(ScannerUpdate):
         'only_notify_superadmin',
         'rule',
         'organization',
-        'contact_person'
+        'contacts'
     ]
 
     def get_success_url(self):
@@ -112,7 +112,7 @@ class GoogleDriveScannerCopy(ScannerCopy):
         'only_notify_superadmin',
         'rule',
         'organization',
-        'contact_person'
+        'contacts'
     ]
 
     def get_initial(self):